Rework tests and add full snapshots into test
